userDiaries renaming...
